Rename linear module
