+++ skfda/ml/regression/linear.py
